+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/qualitynotification/application/investigation/mapper/InvestigationResponseMapper.java
@@ -19,6 +19,7 @@
 
 package org.eclipse.tractusx.traceability.qualitynotification.application.investigation.mapper;
 
+import lombok.experimental.UtilityClass;
 import org.eclipse.tractusx.traceability.common.model.PageResult;
 import org.eclipse.tractusx.traceability.qualitynotification.application.base.mapper.QualityNotificationMapper;
 import org.eclipse.tractusx.traceability.qualitynotification.domain.base.model.QualityNotification;
@@ -36,10 +37,10 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
+import static org.eclipse.tractusx.traceability.qualitynotification.application.base.mapper.QualityNotificationMapper.fromNotifications;
 
+@UtilityClass
 public class InvestigationResponseMapper {
-    private InvestigationResponseMapper() {
-    }
 
     public static InvestigationResponse from(QualityNotification qualityNotification) {
         return InvestigationResponse
@@ -61,7 +62,7 @@ public class InvestigationResponseMapper {
                 .sendToName(getReceiverName(qualityNotification.getNotifications()))
                 .severity(QualityNotificationMapper.from(qualityNotification.getNotifications().stream().findFirst().map(QualityNotificationMessage::getSeverity).orElse(QualityNotificationSeverity.MINOR)))
                 .targetDate(qualityNotification.getNotifications().stream().findFirst().map(QualityNotificationMessage::getTargetDate).map(Instant::toString).orElse(null))
-                .errorMessage(qualityNotification.getErrorMessage())
+                .qualityNotificationMessageResponseList(fromNotifications(qualityNotification.getNotifications()))
                 .build();
     }